pyproject.toml
[project]
name = "dbanalysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.13",
    "python-dotenv>=1.2.2",
]
